Associations added to migrations and mock data added to database

+++ src/dataBase/models/wheelsize.js
@@ -14,7 +14,7 @@ module.exports = (sequelize, DataTypes) => {
     }
   }
   WheelSize.init({
-    number: DataTypes.INTEGER
+    number: DataTypes.DECIMAL
   }, {
     sequelize,
     modelName: 'WheelSize',
